+++ src/components/index.tsx
@@ -11,3 +11,5 @@ export { default as DeleteConfirmAction } from './DeleteConfirmAction';
 export type { ModalActionProps, FormCompPropsConstraint, ModalActionTrigger, ModalActionRef } from './ModalAction';
 export { withDefaultModalActionProps, withModalAction } from './ModalAction';
 export { default as ModalAction } from './ModalAction';
+
+// export * from '../utils';