+++ src/hooks/index.ts
@@ -1,1 +1,5 @@
 export { default as useRefFunction } from './useRefFunction';
+
+export type { ValidatorRuleMap } from './useValidators';
+export type { Validator, RuleRegExpFlags, ValidatorRule } from './useValidatorBuilder';
+export { default as useValidators } from './useValidators';