+++ src/hooks/useContextValidator.ts
@@ -1,8 +1,8 @@
 import { useContext } from 'react';
-import AntHelperContext, { defaultContextValue } from '../components/ConfigProvider/context';
+import ReactEasyContext, { defaultContextValue } from '../components/ConfigProvider/context';
 
-const useValidateContext = () => {
-  const context = useContext(AntHelperContext);
+const useContextValidator = () => {
+  const context = useContext(ReactEasyContext);
   if (!context || context === defaultContextValue) {
     console.warn(
       `ConfigProvider should be wrapped outside of any components of @tiny-codes/react-easy!
@@ -16,4 +16,4 @@ import { ConfigProvider } from '@tiny-codes/react-easy';
   return context;
 };
 
-export default useValidateContext;
+export default useContextValidator;