+++ src/hooks/useLocalizedText.ts
@@ -1,5 +1,5 @@
 import { type ReactNode, useContext } from 'react';
-import AntdHelperContext, { type AntHelperContextProps } from '../components/ConfigProvider/context';
+import AntdHelperContext, { type ReactEasyContextProps } from '../components/ConfigProvider/context';
 
 /**
  * A custom hook that converts input content to localized text using the localize method from
@@ -9,7 +9,7 @@ import AntdHelperContext, { type AntHelperContextProps } from '../components/Con
  *
  * @returns The localized text
  */
-const useLocalizedText: NonNullable<AntHelperContextProps['localize']> = (content, args) => {
+const useLocalizedText: NonNullable<ReactEasyContextProps['localize']> = (content, args) => {
   const { localize } = useContext(AntdHelperContext);
   return localize ? localize(content, args) : (content as ReactNode);
 };